[STYLE] Lint code and separate config from config interfaces
