+++ packages/colors/cli/config.ts
@@ -1,0 +1,11 @@
+export type ThemeFileType = "light" | "dark";
+
+export interface FileConfig {
+  path: string;
+  type: ThemeFileType;
+}
+
+export interface ThemeConfig {
+  output: string;
+  files: FileConfig[];
+}
