+++ packages/colors/cli/gen-themes-cli.ts
