+++ packages/colors/cli/themes.config.ts
@@ -1,14 +1,4 @@
-export type ThemeFileType = "light" | "dark";
-
-export interface FileConfig {
-  path: string;
-  type: ThemeFileType;
-}
-
-export interface ThemeConfig {
-  output: string;
-  files: FileConfig[];
-}
+import {ThemeConfig} from "./config";
 
 const config: ThemeConfig[] = [
   {
